Pull out random utils
