+++ src/api-tests/users.test.ts
@@ -1,56 +1,20 @@
-import range from "lodash/range";
 import random from "lodash/random";
 import knex from "../common/server/knex";
 import { User, NewUser } from "../resources/users/User";
 import { getUsers, createNewUser } from "../pages/api/users";
 import { LadderHistoryItem } from "../resources/users/LadderHistoryItem";
-
-// TODO: move these into util files
-function randomString() {
-  return (
-    Math.random().toString(36).substring(2, 15) +
-    Math.random().toString(36).substring(2, 15)
-  );
-}
-
-function randomBoolean() {
-  return Math.random() >= 0.5;
-}
-
-function randomDate() {
-  return new Date(random(new Date(2012, 1, 1).getTime(), Date.now()));
-}
+import {
+  randomString,
+  randomUser,
+  randomLadderEntry,
+  generateCollection,
+} from "./randomUtils";
 
 const sortByDate = (fieldName: string) => (a: any, b: any) => {
   const dateA = new Date(a[fieldName]);
   const dateB = new Date(b[fieldName]);
   return dateB.getTime() - dateA.getTime();
 };
-
-function generateCollection<T>(
-  supplier: () => Promise<T>,
-  num?: number
-): Promise<T[]> {
-  return Promise.all(range(0, num || random(2, 5)).map(supplier));
-}
-
-async function randomUser(): Promise<User> {
-  const user: Partial<User> = {
-    name: randomString(),
-    archived: randomBoolean(),
-    streak: 0,
-  };
-  return (await knex("users").insert(user).returning("*"))[0];
-}
-
-async function randomLadderEntry(user: User): Promise<LadderHistoryItem> {
-  const ladderEntry: any = {
-    ladder_rung: random(-50, 50),
-    user: user.id,
-    created_at: randomDate(),
-  };
-  return (await knex("ladder_history").insert(ladderEntry).returning("*"))[0];
-}
 
 describe("user endpoints", () => {
   beforeEach(async () => {
@@ -60,7 +24,7 @@ describe("user endpoints", () => {
 
   describe("get users", () => {
     it("should return all unarchived users", async () => {
-      const users = await generateCollection(randomUser);
+      const users = await generateCollection(() => randomUser());
       const expected = users.filter((user) => !user.archived);
       const actual = await getUsers();
 
@@ -74,7 +38,7 @@ describe("user endpoints", () => {
     });
 
     it("should return all the latest ladder entries for each user", async () => {
-      const users = await generateCollection(randomUser);
+      const users = await generateCollection(() => randomUser());
       const usersWithLadder = await Promise.all(
         users.map(async (user) => {
           const ladderEntries = await generateCollection(() =>
